refactor: Rename manager package to core
