+++ src/mg/emberframework/core/data/File.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.data;
+package mg.emberframework.core.data;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
