+++ src/mg/emberframework/core/data/ModelView.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.data;
+package mg.emberframework.core.data;
 
 import java.util.HashMap;
 import java.util.Map;
