+++ src/mg/emberframework/core/data/VerbMethod.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.data;
+package mg.emberframework.core.data;
 
 import java.lang.reflect.Method;
 import java.util.Objects;
