+++ src/mg/emberframework/core/exception/DuplicateUrlException.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.exception;
+package mg.emberframework.core.exception;
 
 public class DuplicateUrlException extends Exception{
     public DuplicateUrlException(String message) {
