+++ src/mg/emberframework/core/exception/IllegalReturnTypeException.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.exception;
+package mg.emberframework.core.exception;
 
 public class IllegalReturnTypeException extends Exception{
     public IllegalReturnTypeException(String message) {
