+++ src/mg/emberframework/core/exception/InvalidRequestException.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.exception;
+package mg.emberframework.core.exception;
 
 public class InvalidRequestException extends Exception{
     public InvalidRequestException(String message) {
