+++ src/mg/emberframework/core/exception/ModelValidationException.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.exception;
+package mg.emberframework.core.exception;
 
 public class ModelValidationException extends Exception{
     public ModelValidationException(String message) {
