+++ src/mg/emberframework/core/exception/UnauthorizedAccessException.java
@@ -1,4 +1,4 @@
-package mg.emberframework.manager.exception;
+package mg.emberframework.core.exception;
 
 public class UnauthorizedAccessException extends Exception {
     public UnauthorizedAccessException(String message) {
