+++ src/mg/emberframework/utils/reflection/ObjectUtils.java
@@ -8,8 +8,8 @@ import java.lang.reflect.Parameter;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import mg.emberframework.annotation.RequestParameter;
-import mg.emberframework.manager.data.File;
-import mg.emberframework.manager.data.Session;
+import mg.emberframework.core.data.File;
+import mg.emberframework.core.data.Session;
 import mg.emberframework.utils.conversion.ObjectConverter;
 import mg.emberframework.utils.io.FileUtils;
