+++ src/mg/emberframework/utils/validation/validators/FieldValidator.java
@@ -3,7 +3,7 @@ package mg.emberframework.utils.validation.validators;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 
-import mg.emberframework.manager.exception.ModelValidationException;
+import mg.emberframework.core.exception.ModelValidationException;
 
 public interface FieldValidator {
     public void validate(String value, Annotation annotation,Field field) throws ModelValidationException;
